Tasklist App/auth/requirements.txt
flask
flask_jwt_extended
mysql-connector-python
python-dotenv
requests
flask_bcrypt